Dynamic site visitor data
n your JavaScript code, Create your own object/array structure to store the number of site visitors each month for the last 6 months. Use that object to dynamically create an HTML table to be displayed on the page. You should be able to change the object to see changes on the website.